+++ k2/src/main/kotlin/at/asitplus/gradle/Testing.kt
@@ -8,7 +8,7 @@ import org.jetbrains.kotlin.gradle.plugin.KotlinDependencyHandler
  * * assertions-core
  * * property
  * * datatest
- * * Testballoon Shim
+ * * Testballoon Addons
  *
  * Also adds `kotlin-reflect` to make kotest work smoothly with IDEA
  */
@@ -20,9 +20,15 @@ inline fun KotlinDependencyHandler.addTestExtensions(target: String? = null) {
     implementation(kotlin("reflect"))
     implementation(project.kotest("assertions-core", target))
     implementation(project.kotest("property", target))
+    implementation("at.asitplus.gradle:testhelper" + (target?.let { "-$it" } ?: "") + ":$buildDate")
     if (System.getProperty("KOTEST_NO_ASP_HELPER") != "true") {
-        implementation("at.asitplus.gradle:testballoon-shim" + (target?.let { "-$it" } ?: "") + ":$buildDate")
         implementation(project.kotest("property"))
         implementation("de.infix.testBalloon:testBalloon-framework-core:${project.AspVersions.testballoon}")
+    }
+    if (System.getProperty("TESTBALLOON_NO_ASP_HELPER") != "true") {
+        implementation("de.infix.testBalloon:testBalloon-framework-core:${project.AspVersions.testballoon}")
+        implementation("at.asitplus.testballoon:fixturegen-freespec:${project.AspVersions.testballoonAddons}")
+        implementation("at.asitplus.testballoon:datatest:${project.AspVersions.testballoonAddons}")
+        implementation("at.asitplus.testballoon:property:${project.AspVersions.testballoonAddons}")
     }
 }