+++ CHANGELOG.md
@@ -3,9 +3,13 @@
 Starting with May 2025, the conventions plugin does not include a Kotlin version anymore. Instead, it lets any consuming
 project set it, just as with the Android Gradle Plugin. **This is a breaking change, as there is no fallback version specified.**
 
+## 20250728
+* Fix Kotest KSP wiring for good
+
 ## 20250727
 * Kotest 6.0.0.M6
 * Removed https://s01.oss.sonatype.org/content/repositories/snapshots from maven repos
+* `XmlReportingProjectConfig` Kotest abstract project config shim to enable KMP JUnit XML reporting
 
 ## 20250723+1
 * fix order of XCF export arguments