+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 Starting with May 2025, the conventions plugin does not include a Kotlin version anymore. Instead, it lets any consuming
 project set it, just as with the Android Gradle Plugin. **This is a breaking change, as there is no fallback version specified.**
 
-## TBA
+## 20251107
 * Allow nested typealias
 * TestBalloon 0.7 support